Move extlib/ to vint/_bundles/
