+++ setup.cfg
@@ -3,5 +3,5 @@ description-file = README.markdown
 
 [flake8]
 max-line-length = 100
-exclude = extlib/*
+exclude = vint/_bundles/*
 ignore = E303
